Removed logic where driver buffer size is set at open. Now it is checked. Also added newer error codes for buffer size related issues
Previous commit added logic such that when raspa open was called with a specified buffer size as arg,
raspa would write to the audio_buffer_size param file and set the buffer size. However, since not all
drivers support this, it was decided that raspa will revert back to the logic of checking if the buffer
size matches with the driver. The setting of the driver buffer size is now handled externally

the logic to propogate an error code in case the driver has been configured with an invalid buffer size
is still retained. this is to prevent scenarios where both raspa app and the driver are configured
for an invalid buffer size.